Perform graceful stops on PHP-FPM
The default signal used by docker to stop a container is SIGTERM and
that signal is understood by PHP-FPM as immediate termination.

This means that `php-fpm` won't wait for in-flight requests to finish
before of stopping the workers and main process. This behaviour is quite
undesired for production environments and should be avoided as much as
possible.

This implementation is meant to be extremely simple and only address the
stop signal used when running `docker stop`. `php-fpm` signal choices
are somewhat peculiar (`SIGUSR2` for example) and if we want to "fix"
them we would have to use a tool (e.g.: `dumb-init`).

More info:

 - https://linux.die.net/man/8/php-fpm
 - https://docs.docker.com/engine/reference/builder/#stopsignal
 - https://github.com/Yelp/dumb-init
 - https://github.com/usabilla/php-docker-template/pull/102